// projects/adminmodule/src/app/models/responsemodel.ts
  export interface Ires_RollCalllist {
  FullName: string;
  Aadhaar: string;
  Rollno: string;
  Batch_division: string;
  MobileNumber: number;
  EmailID: string;
  Subject_Group_Code: string;
  Billdeskdate: string;
  Prnno: string;
  Emailid: string;
  Annual_income: string;
}